+++ src/components/inventory.ts
@@ -15,6 +15,16 @@ export class Inventory {
 		this.capacity = capacity
 	}
 		
+	clone(newParent: Actor) {
+		let newInventory = new Inventory(this.engine, newParent, this.capacity)
+
+		for (let item of this.items) {
+			newInventory.items.add(item.clone(newInventory))
+		}
+
+		return newInventory
+	}
+
 
 	/** Removes an item from the inventory and restores it to the game map, at the player's current location */
 	drop(item: Item): void {