+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - script to generate map from image
 - initial checks on screen resolution and touch devices
 - configurable start player position in maps
+- actor loader
 
 ## [0.1.2] - 2021-02-02
 ### Fixed